a1-image: Add wifi and wwan support to NetworkManager
In order for the NetworkManager to be able to build with wwan support,
it also needs to connect to the Modemmanager. Therefore the Modemmanager
is selected as a NetworkManager option.

Furthermore, the p2p feature of the rsi9113 seems to be broken. Therefore
we disable p2p as well as randomized mac for scanning in the NetworkManager
config.