networkmanager: Add configurable default connections
The end0/1 connections are normal dhcp client connections, which connect
to any given dhcp server.

The end0/1-dhcp connections set up a shared network. This starts a dhcp server
on the corresponding port with host address 10.0.0.1 and 10.0.1.1 respectively.

Per default the dhcp connections don't autoconnect, except if END0_DHCP or END1_DHCP
are set to true or 1.

The priority of the connections is set to -20 for the normal mode and -10 for shared
connections. Therefore per default any newly generated connection will have higher
priority. Furthermore the dhcp connection have higher priority then the normal ones,
if they are set to autoload.