Dynamic-Layers: Add a1-ubi-image append files
Sadly the 'require a1-image.bb' doesn't pick up on the bbappend files. Therefore
we duplicate the bbappend files in the dynamic layers. This has the added advantage
that we can customize them seperately. For example, pip is not going to be needed in
the ubi image, as the rootfs is read-only which makes the installation into the
default site-packages impossible. We might however add this back later and install
the packages in the .local folder