python: fixed python wrongly being added into image
Just adding the python dependencies to the image is unwise, as some layers
depend on on the python layer for host-side computation but don't need it
at runtime. Therefore we explicitly check in here what config has been set.
This allows also for extending this in the future for a finer controll over
additionally installed packages.